HEPExampleProject

Documentation for HEPExampleProject.

HEPExampleProject.ALPHA

HEPExampleProject.ELECTRON_MASS

HEPExampleProject.MUON_MASS

HEPExampleProject.Event

HEPExampleProject.FourMomentum

Base.:*

Base.:+

Base.:-

HEPExampleProject.build_unweighting_mask

HEPExampleProject.coords_to_dict

HEPExampleProject.differential_cross_section

HEPExampleProject.filter_accepted

HEPExampleProject.generate_event_and_masks

HEPExampleProject.generate_events

HEPExampleProject.generate_events_cpu

HEPExampleProject.generate_flat_event

HEPExampleProject.generate_flat_events_cpu

HEPExampleProject.max_weight

HEPExampleProject.minkowski_dot

HEPExampleProject.total_cross_section

HEPExampleProject.ALPHA — Constant

ALPHA

The fine-structure constant at tree level, given as 0.0072973525643(11).

This value is taken from the PDG 2020 recommended values of the fundamental physical constants.

References

The Review of Particle Physics: P.A. Zyla et al. (Particle Data Group), Prog. Theor. Exp. Phys. 2020, 083C01 (2020)

HEPExampleProject.ELECTRON_MASS — Constant

ELECTRON_MASS

The mass of the electron in its rest frame, given as 0.51099895000 MeV.

This value is taken from the PDG 2020 recommended values of the fundamental physical constants.

References

The Review of Particle Physics: P.A. Zyla et al. (Particle Data Group), Prog. Theor. Exp. Phys. 2020, 083C01 (2020)

HEPExampleProject.MUON_MASS — Constant

MUON_MASS

The mass of the muon in its rest frame, given as 105.6583755 MeV.

This value is taken from the PDG 2020 recommended values of the fundamental physical constants.

References

The Review of Particle Physics: P.A. Zyla et al. (Particle Data Group), Prog. Theor. Exp. Phys. 2020, 083C01 (2020)

HEPExampleProject.Event — Type

Event(
    electron_momentum,
    positron_momentum,
    muon_momentum,
    anti_muon_momentum,
    weight
)

Represents a single event in the process $e^+ e^- \to \mu^+ \mu^-$, containing the four-momenta of the electron, and positron in the initial state and muon, and anti-muon in the final state, along with the corresponding event weight.

Fields

electron_momentum::FourMomentum: The four-momentum of the incoming electron.

positron_momentum::FourMomentum: The four-momentum of the incoming positron.

muon_momentum::FourMomentum: The four-momentum of the outgoing muon.

anti_muon_momentum::FourMomentum: The four-momentum of the outgoing anti-muon.

weight::T: The weight of the event, representing the likelihood of the event occurring. Must be non-negative.

Example

julia> p_electron = FourMomentum(10.0, 0.0, 0.0, 9.0)
FourMomentum(en = 10.0, x = 0.0, y = 0.0, z = 9.0)

julia> p_positron = FourMomentum(10.0, 0.0, 0.0, -9.0)
FourMomentum(en = 10.0, x = 0.0, y = 0.0, z = -9.0)

julia> p_muon = FourMomentum(9.5, 2.0, 1.0, 8.0)
FourMomentum(en = 9.5, x = 2.0, y = 1.0, z = 8.0)

julia> p_anti_muon = FourMomentum(10.5, -2.0, -1.0, -8.0)
FourMomentum(en = 10.5, x = -2.0, y = -1.0, z = -8.0)

julia> weight = 1.2
1.2

julia> event = Event(p_electron, p_positron, p_muon, p_anti_muon, weight)
Event e-e+ -> mu-mu+
	electron:  (10.0, 0.0, 0.0, 9.0)

	positron:  (10.0, 0.0, 0.0, -9.0)

	muon:      (9.5, 2.0, 1.0, 8.0)

	anti-muon: (10.5, -2.0, -1.0, -8.0)

	weight:    1.2

HEPExampleProject.FourMomentum — Type

FourMomentum{T<:Real}(en::T, x::T, y::T, z::T)

Defines a four-momentum vector energy and cartesian spatial components, where the type T is a subtype of Real, allowing for arbitrary precision and numeric types.

Fields

en: Energy component.

x: Spatial component in the x-direction.

y: Spatial component in the y-direction.

z: Spatial component in the z-direction.

Example

julia> FourMomentum(4.0, 1.0, 2.0, 3.0)
FourMomentum(en = 4.0, x = 1.0, y = 2.0, z = 3.0)

julia> FourMomentum(4, 1.0, 2, 3) # implicit type promotion
FourMomentum(en = 4.0, x = 1.0, y = 2.0, z = 3.0)

Base.:* — Method

Base.:*(a::Real,p2::FourMomentum)

Defines scalar multiplication for a FourMomentum object, scaling each component of the four-momentum by a scalar a.

Example

julia> p = FourMomentum(4.0, 1.0, 2.0, 3.0)
FourMomentum(en = 4.0, x = 1.0, y = 2.0, z = 3.0)

julia> 2 * p
FourMomentum(en = 8.0, x = 2.0, y = 4.0, z = 6.0)

Base.:+ — Method

Base.:+(p1::FourMomentum,p2::FourMomentum)

Defines vector addition for two FourMomentum objects. The result is a new FourMomentum with each component being the sum of the corresponding components of p1 and p2.

Example

julia> p1 = FourMomentum(4.0, 1.0, 2.0, 3.0)
FourMomentum(en = 4.0, x = 1.0, y = 2.0, z = 3.0)

julia> p2 = FourMomentum(2.0, 0.5, 1.0, 1.5)
FourMomentum(en = 2.0, x = 0.5, y = 1.0, z = 1.5)

julia> p1 + p2
FourMomentum(en = 6.0, x = 1.5, y = 3.0, z = 4.5)

Base.:- — Method

Base.:-(p1::FourMomentum,p2::FourMomentum)

Defines vector subtraction for two FourMomentum objects. The result is a new FourMomentum with each component being the difference between the corresponding components of p1 and p2.

Example

julia> p1 = FourMomentum(4.0, 1.0, 2.0, 3.0)
FourMomentum(en = 4.0, x = 1.0, y = 2.0, z = 3.0)

julia> p2 = FourMomentum(2.0, 0.5, 1.0, 1.5)
FourMomentum(en = 2.0, x = 0.5, y = 1.0, z = 1.5)

julia> p1 - p2
FourMomentum(en = 2.0, x = 0.5, y = 1.0, z = 1.5)

HEPExampleProject.build_unweighting_mask — Method

build_unweighting_mask(E_in::T, event) where T

Build a mask to unweight events using the rejection sampling method. The mask filters events based on their weights, comparing each event's weight against a random value scaled by the maximum weight for the given incoming energy E_in.

Arguments

E_in::T: Incoming energy of the event (must be a subtype of Real).

event: The Event object whose weight will be compared.

Returns

A Boolean value indicating whether the event passes the unweighting criterion (true if accepted, false otherwise).

HEPExampleProject.coords_to_dict — Method

coords_to_dict(E_in::Real, cos_theta::Real, phi::Real)

Constructs the four-momenta for an electron-positron annihilation process $e^+ e^- \rightarrow \mu^+ \mu^-$ in the center-of-mass frame. The input energy (E_in), scattering angle (cos_theta), and azimuthal angle (phi) are used to compute the incoming and outgoing momenta for the particles.

Returns

A Dict mapping the particle names ("e-", "e+", "mu-", "mu+") to their respective FourMomentum objects.

Example

julia> mom_dict = coords_to_dict(1e3,0.9,pi/4)
Dict{String, FourMomentum{Float64}} with 4 entries:
  "mu+" => (1000.0, -306.495431, -306.495431, -894.962239)…
  "mu-" => (1000.0, 306.495431, 306.495431, 894.962239)…
  "e+"  => (1000.0, 0.0, 0.0, -999.999869)…
  "e-"  => (1000.0, 0.0, 0.0, 999.999869)…

julia> mom_dict["e-"]
FourMomentum(en = 1000.0, x = 0.0, y = 0.0, z = 999.999869440028)

julia> mom_dict["e+"]
FourMomentum(en = 1000.0, x = 0.0, y = 0.0, z = -999.999869440028)

julia> mom_dict["mu-"]
FourMomentum(en = 1000.0, x = 306.4954310103767, y = 306.49543101037665, z = 894.9622389946002)

julia> mom_dict["mu+"]
FourMomentum(en = 1000.0, x = -306.4954310103767, y = -306.49543101037665, z = -894.9622389946002)

HEPExampleProject.differential_cross_section — Method

differential_cross_section(E_in::Real, cos_theta::Real)

Calculates the differential cross section for the process $e^+ e^- \to \mu^+ \mu^-$ at tree level in quantum electrodynamics (QED). The calculation is performed in the center-of-momentum frame (CMS) as a function of the energy (E_in) of the incoming electron and the cosine of the scattering angle (cos_theta), i.e. the angle between the incoming electron and outgoing muon.

Arguments

E_in::Real: The energy of the electron/positron in MeV. In the center-of-momentum frame, this energy is equal to the energies of all incoming and outgoing particles.

cos_theta::Real: The cosine of the scattering angle between the incoming electron and the outgoing muon.

Methodology

This function computes the differential cross section using the following formula (from Schwartz 2014):

\[\frac{\mathrm{d}\sigma}{\mathrm{d}\Omega} = \frac{\alpha^2}{16 E_{\mathrm{in}}^6}\left( E_{\text{in}}^4 + \rho_e^2 \rho_\mu^2 \cos^2\theta + E_{\text{in}}^2 \left( m_e^2 + m_\mu^2 \right) \right)\]

where:

$E_{\text{in}}$ is the energy of the incoming electron/positron in CMS,

$\rho_i = \sqrt{E_{\text{in}}^2 - m_i^2}$ are the magnitude of three-momenta of the electron ($i=e$) and muon ($i=\mu$) in the center-of-mass frame,

$\alpha$ is the fine-structure constant,

$m_e$ and $m_\mu$ are the masses of the electron and muon, respectively.

Example

julia> E_in = 1e3 # MeV
1000.0

julia> cos_theta = 0.5
0.5

julia> differential_cross_section(E_in, cos_theta)
4.164686491998452e-12

References

Schwartz 2014: M.D. Schwartz, "Quantum Field Theory and the Standard Model", Cambridge University Press, New York (2014)

HEPExampleProject.filter_accepted — Method

filter_accepted(record_list)

Filter out and return only the accepted events from a list of event records. Each record contains an event and a Boolean mask indicating whether the event was accepted.

Arguments

record_list: A list of tuples where each tuple contains an Event and a corresponding Boolean mask.

Returns

A list of accepted Event objects.

HEPExampleProject.generate_event_and_masks — Method

generate_event_and_masks(E_in)

Generate an event along with its unweighting mask. The mask is built based on the event's weight compared to a randomly scaled maximum weight (see build_unweighting_mask for details).

Arguments

E_in: Energy of the incoming electron (can be any Real type).

Returns

A tuple containing:

event: The generated event.

mask: A Boolean indicating whether the event is accepted or rejected based on its weight.

HEPExampleProject.generate_events — Method

generate_events(E_in::T, nevents; array_type::Type{ARRAY_TYPE}=Vector{T}, chunksize=100) 
where {T<:Real, ARRAY_TYPE<:AbstractVector{T}}

Generate a specified number of unweighted events using rejection sampling. Events are generated in chunks, and only accepted events are retained.

Arguments

E_in::T: Energy of the incoming electron (must be a subtype of Real).

nevents: The number of unweighted events to generate.

array_type::Type{ARRAY_TYPE}: Optional; the type of array to use for the internal event generation (default is Vector{T}).

chunksize: Optional; the number of events to generate per chunk (default is 100).

Returns

A list of unweighted Event objects.

HEPExampleProject.generate_events_cpu — Method

generate_events_cpu(rng::AbstractRNG, E_in::Real, nevents::Int)

Generates a list of unweighted events for the process $e^+ e^- \to \mu^+ \mu^-$ using the acceptance-rejection method. The events are generated according to the differential_cross_section.

Arguments

rng::AbstractRNG: A random number generator instance, used for sampling random numbers.

E_in::Real: The center-of-mass energy of the incoming particles.

nevents::Int: The number of unweighted events to generate.

Returns

unweighted_events: A list of Event objects.

Methodology

The function generates trial events by randomly sampling cos_theta and calculating the event weight from the differential cross section.

Events are accepted if the sampled weight passes an acceptance-rejection test, where the probability of acceptance is proportional to the event weight.

The azimuthal angle (phi) is randomly sampled for accepted events, and the corresponding four-momenta for the incoming electron and positron, as well as the outgoing muons are constructed.

This process repeats until the desired number of events (nevents) is generated.

Example

julia> rng = MersenneTwister(137)
julia> unweighted_events = generate_events_cpu(rng, 1e3, 1000);

HEPExampleProject.generate_flat_event — Method

generate_flat_event(E_in::T) where {T<:Real}

Generate a flat event with a given incoming energy E_in. The event is characterized by a randomly sampled cosine of the polar angle (cth), azimuthal angle (phi), and a weight. The weight is computed using the differential cross-section for the process.

Arguments

E_in::T: Incoming energy of the event (must be a subtype of Real).

Returns

An Event object containing the generated event's properties.

HEPExampleProject.generate_flat_events_cpu — Method

generate_flat_events_cpu(rng::AbstractRNG, E_in::Real, nevents::Int)

Generates a list of weighted events for the process $e^+ e^- \to \mu^+ \mu^-$ using a flat random distribution in the cosine of the scattering angle and azimuthal angle.

Arguments

rng::AbstractRNG: A random number generator instance, used for sampling random numbers.

E_in::Real: The energy of the incoming electron.

nevents::Int: The number of events to generate.

Returns

event_list: A list of Event objects, where each event contains:

Methodology

Randomly generates the cosine of the scattering angle (cos_theta) and azimuthal angle (phi) for each event.

Uses the differential_cross_section to compute the weights for each event.

Constructs the four-momenta of the outgoing muons and anti-muons from the random angles and the input energy E_in.

Returns a list of Event objects containing the momenta and corresponding event weights.

Example

julia> rng = MersenneTwister(137)
julia> event_list = generate_flat_events_cpu(rng, 1e3, 1000);

The ; should be used at the end of the last prompt to suppress printing the whole event list.

Notes

This method generates weighted events where the weights are derived from the differential cross section. For unweighted events, use generate_events_cpu.

HEPExampleProject.max_weight — Method

max_weight(E_in)

Return the maximum weight, i.e. the maximum value of differential_cross_section for specified initial electron energy E_in.

Example

julia> max_weight(1e3)
6.619160791345756e-12

HEPExampleProject.minkowski_dot — Method

minkowski_dot(p1::FourMomentum, p2::FourMomentum)

Computes the Minkowski dot product of two four-momentum vectors. The dot product uses the Minkowski metric (+,-,-,-). For $p_i = (E_i,p_i^x,p_i^y,p_i^z)$ with $i=1,2$, the result is:

\[ p_1 \cdot p_2 = E_1E_2 - p_1^xp_2^x - p_1^yp_2^y - p_1^zp_2^z\]

Example

julia> p1 = FourMomentum(4.0, 1.0, 2.0, 3.0)
FourMomentum(en = 4.0, x = 1.0, y = 2.0, z = 3.0)

julia> p2 = FourMomentum(3.0, 0.5, 1.0, 1.5)
FourMomentum(en = 3.0, x = 0.5, y = 1.0, z = 1.5)

julia> minkowski_dot(p1, p2)
5.0

HEPExampleProject.total_cross_section — Method

total_cross_section(E_in::Real)

Calculates the total cross section for the process $e^+ e^- \to \mu^+ \mu^-$ at tree level, as a function of the initial electron energy E_in.

Arguments

E_in::Real: The energy of the initial electron/positron in MeV.

Methodology

The calculation is based on the analytical integration of the differential_cross_section and uses the following formula for the total cross section:

\[\mathrm{d}\sigma(E_in) = \frac{\pi \alpha^2}{8 E_{\text{in}}^6} \cdot \frac{\rho_\mu}{\rho_e} \left( 2 E_{\text{in}}^4 + \frac{2}{3} \rho_\mu^2 \rho_e^2 + 2 E_{\text{in}}^2 (m_\mu^2 + m_e^2) \right)\]

where:

$E_{\text{in}}$ is the energy of the incoming electron/positron in CMS,

$\rho_i = \sqrt{E_{\text{in}}^2 - m_i^2}$ are the magnitude of three-momenta of the electron ($i=e$) and muon ($i=\mu$) in the center-of-mass frame,

$\alpha$ is the fine-structure constant,

$m_e$ and $m_\mu$ are the masses of the electron and muon, respectively.

Example

julia> E_in = 1e3 # MeV
1000.0

julia> total_cross_section(E_in)
5.576208658540326e-11